refactor(messages): lifted the message state up App.js so that both the chatsession and chatcontainer elements can modify
